[Silver I] Title: 쉬운 계단 수, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Silver/10844. 쉬운 계단 수/쉬운 계단 수.cc
@@ -2,8 +2,8 @@
 
 using namespace std;
 typedef long long ll;
-ll d[104][12]; //오큰수의 갯수 [길이][시작숫자]
-ll n;
+int d[104][12]; //오큰수의 갯수 [길이][시작숫자]
+int n;
 
 int main() {
 	ios::sync_with_stdio(0);
@@ -14,12 +14,12 @@ int main() {
 
 	for (int i = 2; i <= 100; ++i) {
 		for (int j = 1; j <= 10; ++j) {
-			d[i][j] = (d[i-1][j-1] % 1000000000 +d[i-1][j+1] % 1000000000)%1000000000;
+			d[i][j] = (d[i-1][j-1]%1000000000+d[i-1][j+1]%1000000000)%1000000000;
 		}
 	}
 
 	ll ret = 0;
-	for (int j = 0; j <= 10; ++j) ret = (ret % 1000000000 + d[n][j] % 1000000000)% 1000000000;
+	for (int j = 0; j <= 10; ++j)ret = (ret % 1000000000 + d[n][j] % 1000000000)% 1000000000;
 	cout << ret;
 
 	return 0;
